move explorerActionManager & explorerActionState to shared
