+++ plugins/amazonq/shared/jetbrains-community/resources/META-INF/module-amazonq.xml
@@ -7,6 +7,6 @@
     <xi:include href="/META-INF/plugin-codewhisperer.xml" />
 
     <extensions defaultExtensionNs="com.intellij">
-        <applicationService serviceImplementation="migration.software.aws.toolkits.jetbrains.services.codewhisperer.explorer.CodeWhispererExplorerActionManager"/>
+        <applicationService serviceImplementation="software.aws.toolkits.jetbrains.services.codewhisperer.CodeWhispererExplorerActionManager"/>
     </extensions>
 </idea-plugin>
