+++ plugins/amazonq/shared/jetbrains-community/src/software/aws/toolkits/jetbrains/services/codewhisperer/CodeWhispererLoginType.kt
@@ -1,7 +1,7 @@
 // Copyright 2022 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 // SPDX-License-Identifier: Apache-2.0
 
-package software.aws.toolkits.jetbrains.services.codewhisperer.credentials
+package software.aws.toolkits.jetbrains.services.codewhisperer
 
 enum class CodeWhispererLoginType(val displayName: String) {
     SSO("IAM Identity Center"),
